% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircleArc.R
\name{CircleArcEquate}
\alias{CircleArcEquate}
\title{Circle-arc based on two vectors of data assuming from equivalent groups}
\usage{
CircleArcEquate(xscores, yscores, maxX, maxY, minX = 0, minY = 0)
}
\arguments{
\item{xscores}{A vector of scores from individual candidates on form X.}

\item{yscores}{A vector of scores from individual candidates on form Y.}

\item{maxX}{Maximum score on form X (or upper point for circle arc method).}

\item{maxY}{Maximum score on form Y (or upper point for circle arc method).}

\item{minX}{Minimum score on form X (or lower point for circle arc method). Default of zero.}

\item{minY}{Minimum score on form Y (or lower point for circle arc method). Default of zero.}
}
\value{
The function returns a list with the following elements:
\describe{
  \item{yxFunc}{A function that translates any vector of scores on form X into equivalent scores on form Y.}
  \item{EqTable}{A data frame showing the equivalent score on form Y for every integer score between minX and maxxX on form X.}
}
}
\description{
Calculates means an applies .
}
\examples{
x=round(runif(500,0,50))
y=round(pmin(50,pmax(0,rnorm(500,35,10))))
circ_eq=CircleArcEquate(x,y,50,50)
plot(circ_eq$EqTable$x,circ_eq$EqTable$equiyx,type='l')
circ_eq$EqTable

}
