% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircleArc.R
\name{CircleArcFromMeans}
\alias{CircleArcFromMeans}
\title{Find circle-arc equating functions based on minima, maxima and means for two tests}
\usage{
CircleArcFromMeans(meanX, meanY, minX, minY, maxX, maxY)
}
\arguments{
\item{meanX}{Mean score on form X (or any midpoint for circle arc method).}

\item{meanY}{Mean score on form Y (or any midpoint for circle arc method).}

\item{minX}{Minimum score on form X (or lower point for circle arc method).}

\item{minY}{Minimum score on form Y (or lower point for circle arc method).}

\item{maxX}{Maximum score on form X (or upper point for circle arc method).}

\item{maxY}{Maximum score on form Y (or upper point for circle arc method).}
}
\value{
The function returns a list with the following elements:
\describe{
  \item{yxFunc}{A function that translates any vector of scores on form X into equivalent scores on form Y.}
  \item{EqTable}{A data frame showing the equivalent score on form Y for every integer score between minX and maxxX on form X.}
}
}
\description{
Just apply the standard circle-arc equating functions based on the three points on the arc for each form.
}
\examples{
circ_eq=CircleArcFromMeans(20, 30, 0, 0, 50, 50)
plot(circ_eq$EqTable$x,circ_eq$EqTable$equiyx,type='l')
points(c(0,20,50),c(0,30,50))#show the three points

}
