% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel Equating Functions.R
\name{DistToMuSig}
\alias{DistToMuSig}
\title{Function to calculate mu and sigma from a score distribution}
\usage{
DistToMuSig(scores, count)
}
\arguments{
\item{scores}{Vector of possible scores}

\item{count}{Number of people achieving each score (weighted numbers, probabilities or proportions also acceptable). Must be same length as scores.}
}
\value{
The function returns a list with the following elements:
\describe{
  \item{mu}{The mean.}
  \item{sigma}{The standard deviation.}
}
}
\description{
Function to calculate mu and sigma from a score distribution
}
\examples{
x=pmax(0,pmin(10,round(rnorm(200,5,2))))
counts=200*(ecdf(x)(0:10)-ecdf(x)(-1:9))#used rather than "table" so all scores included even if zero count
DistToMuSig(0:10,counts)

}
\keyword{KernEqWPS}
