% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNweights.R
\docType{data}
\name{EquateNNWeights}
\alias{EquateNNWeights}
\title{A list of weights and biases for experimental equating using neural networks}
\format{An object of class \code{list} of length 12.}
\usage{
EquateNNWeights
}
\description{
This list forms the basis of the \emph{PredPercentileNN} and \emph{EquateNN} functions. It is mainly supplied to satisfy any 
curiosity about the way in which these functions work.
The weights and biases in this list are unlikely to be of real interest outside of the functions that 
utilise them.
}
\keyword{datasets}
