% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel Equating Functions.R
\name{Fhx}
\alias{Fhx}
\title{Continuized cumulative density function from a score distribution for a given bandwidth}
\usage{
Fhx(x, fX, scores, mu, sigma, h)
}
\arguments{
\item{x}{Score(s) for which we wish to calculate the density}

\item{fX}{Number of people achieving each score (weighted numbers, probabilities or proportions also acceptable). Must be same length as scores.}

\item{scores}{Vector of possible scores.}

\item{mu}{Mean score in distribution.}

\item{sigma}{Standard deviation score in distribution.}

\item{h}{Bandwidth used for continuization.}
}
\description{
Note: mu and sigma are parameters in here (rather than being calculated from distribution) 
to avoid repitative recalculation when applying this function within the routine to optimise bandwidths
}
\keyword{KernEqWPS}
