% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel Equating Functions.R
\name{FindAVDBandwidth}
\alias{FindAVDBandwidth}
\title{Method to estimate appropriate bandwidth using the Andersson and von Davier plug-in formula.}
\usage{
FindAVDBandwidth(scores, fX)
}
\arguments{
\item{scores}{Vector of possible scores.}

\item{fX}{Number of people achieving each score. Weighted numbers are also acceptable but must add up to the overall sample size. Must be same length as scores.}
}
\value{
The function returns a single value representing the suggested bandwidth.
}
\description{
The formula for suggesting an appropriate bandwidth is very simple and can easily be seen by looking at the function code.
}
\examples{
#EXAMPLE OF FINDING BANDWIDTH TO EQUATE FIRST AND SECOND HALVES OF MATHS TEST
x1=rowSums(mathsdata[,1:62])
y1=rowSums(mathsdata[,63:124])
summary(x1)
summary(y1)
tabx1=table(x1)
taby1=table(y1)

RoughHx=FindAVDBandwidth(as.numeric(names(tabx1)),as.vector(tabx1))
RoughHx

RoughHy=FindAVDBandwidth(as.numeric(names(taby1)),as.vector(taby1))
RoughHy

eq1=KernelEquateFromScoresEG(x1,y1,hX=RoughHx,hY=RoughHy)
plot(sort(unique(x1)),eq1$yx,type='l')

#prove that above code is the same as the default method of equating
eq2=KernelEquateFromScoresEG(x1,y1)
lines(sort(unique(x1)),eq2$yx,lty=2,col="purple")
}
\references{
Andersson, B., & von Davier, A. A. (2014). Improving the bandwidth selection in kernel equating. 
\emph{Journal of Educational Measurement, 51}(3), 223-238.
}
\keyword{KernEqWPS}
