% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel Equating Functions.R
\name{FindBestBandwidth1Iter}
\alias{FindBestBandwidth1Iter}
\title{One iteration of method to find optimal bandwidth (between 0.05 and 30) that minimses cross-validation error between two half-samples.}
\usage{
FindBestBandwidth1Iter(scores, fX)
}
\arguments{
\item{scores}{Vector of possible scores.}

\item{fX}{Number of people achieving each score (should be whole numbers). Must be same length as scores.}
}
\value{
The function returns a single value representing the optimal bandwidth.
}
\description{
Based on paper by Liang and von Davier (2014). It works by splitting the data set in half, fitting a continuized distribution with a gaussian kernel 
to one half of the data, and then evaluating the log-likelihood of this distribution against the other half of the data.
}
\examples{
tot1=rowSums(mathsdata)
tab1=table(tot1)
FindBestBandwidth1Iter(as.numeric(names(tab1)),as.vector(tab1))
}
\references{
Liang, T., & von Davier, A. A. (2014). Cross-validation: An alternative bandwidth-selection method in kernel equating. 
\emph{Applied Psychological Measurement, 38}(4), 281-295.
}
\keyword{KernEqWPS}
