% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel Equating Functions.R
\name{KernelChainedEquate}
\alias{KernelChainedEquate}
\title{Apply kernel chained equating}
\usage{
KernelChainedEquate(dx, dy, hX = NA, hY = NA, hA = NA)
}
\arguments{
\item{dx}{Data frame with variables "x" and "a" representing scores for individual candidates on form X and on the anchor test.}

\item{dy}{Data frame with variables "y" and "a" representing scores for individual candidates on form Y and on the anchor test.}

\item{hX}{Bandwidth for form X. By default a plug-in estimator from Andersson and von Davier (2014) is used.}

\item{hY}{Bandwidth for form Y. By default a plug-in estimator from Andersson and von Davier (2014) is used.}

\item{hA}{Bandwidth for the anchor form. By default a plug-in estimator from Andersson and von Davier (2014) is used which will differ across populations. If a value is specified then the same bandwidth will be used in both populations.}
}
\value{
The function returns a list with the following elements:
\describe{
  \item{chainedFunc}{A function that translates any vector of scores on form X into equivalent scores on form Y.}
  \item{EqTable}{A data frame combining the sorted unique scores on form X in the data and their equated values on form Y (chained kernel).}
}
}
\description{
Chained equating via an anchor test is done by successively applying the \emph{KernelEquateFromScoresEG} function between.
}
\examples{
#Simulate two data sets with roughly equivalent relationship to underlying "true" scores but a difference in means
n1=8000
n2=5500
t1=rnorm(n1,0.5,1)
t2=rnorm(n2,0,1)
x=round(pmin(100,pmax(0,50+20*(0.9*t1+rnorm(n1,0,sqrt(1-0.9^2))))))
a1=round(pmin(20,pmax(0,10+4*(0.7*t1+rnorm(n1,0,sqrt(1-0.7^2))))))
cor(cbind(x,t1,a1))
y=round(pmin(100,pmax(0,50+20*(0.9*t2+rnorm(n2,0,sqrt(1-0.9^2))))))
a2=round(pmin(20,pmax(0,10+4*(0.7*t2+rnorm(n2,0,sqrt(1-0.7^2))))))
cor(cbind(y,t2,a2))
KernelChainedEquate(data.frame(x=x,a=a1),data.frame(y=y,a=a2))

}
\references{
Andersson, B., & von Davier, A. A. (2014). Improving the bandwidth selection in kernel equating. 
\emph{Journal of Educational Measurement, 51}(3), 223-238.
}
\keyword{KernEqWPS}
