% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel Equating Functions.R
\name{KernelEquateFromDists}
\alias{KernelEquateFromDists}
\title{Apply Kernel Equating to Equivalent Groups.}
\usage{
KernelEquateFromDists(scoresX, fX, scoresY, fY, hX, hY)
}
\arguments{
\item{scoresX}{Vector of possible scores on form X.}

\item{fX}{Number of people achieving each score on form X (weighted numbers, probabilities or proportions also acceptable). Must be same length as scoresX.}

\item{scoresY}{Vector of possible scores on form Y.}

\item{fY}{Number of people achieving each score on form Y (weighted numbers, probabilities or proportions also acceptable). Must be same length as scoresY.}

\item{hX}{Bandwidth for form X.}

\item{hY}{Bandwidth for form Y.}
}
\value{
The function returns a list with the following elements:
\describe{
  \item{yxFunc}{A function that translates any vector of scores on form X into equivalent scores on form Y.}
  \item{yx}{A vector representing the scores derived from applying yxFunc to scoresX.}
}
}
\description{
Assumes that X and Y are taken by equivalent groups (or by the same group). Requires bandwidths for continuization to be supplied.
}
\keyword{KernEqWPS}
