% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel Equating Functions.R
\name{KernelEquateFromScoresEG}
\alias{KernelEquateFromScoresEG}
\title{Apply kernel equating based on two vectors of scores on two forms from equivalent candidates.}
\usage{
KernelEquateFromScoresEG(xscores, yscores, hX = NA, hY = NA)
}
\arguments{
\item{xscores}{A vector of scores from individual candidates on form X.}

\item{yscores}{A vector of scores from individual candidates on form Y.}

\item{hX}{Bandwidth for form X. By default a plug-in estimator from Andersson and von Davier (2014) is used.}

\item{hY}{Bandwidth for form Y. By default a plug-in estimator from Andersson and von Davier (2014) is used.}
}
\value{
The function returns a list with the following elements:
\describe{
  \item{yxFunc}{A function that translates any vector of scores on form X into equivalent scores on form Y.}
  \item{yx}{A vector representing the scores derived from applying yxFunc to a sorted list of all the unqiue form X scores that actually occur in the data.}
}
}
\description{
Apply kernel equating based on two vectors of scores on two forms from equivalent candidates.
}
\examples{
#simple example of equating a roughly normally distributed sets of scores to one with a roughly logistic distribution
x=pmax(0,pmin(100,round(rnorm(300,50,15))))
y=pmax(0,pmin(100,round(rlogis(300,60,16))))
hist(x)
hist(y)
KE1=KernelEquateFromScoresEG(x,y)
plot(0:100,KE1$yxFunc(0:100),type='l')

}
\references{
Andersson, B., & von Davier, A. A. (2014). Improving the bandwidth selection in kernel equating. 
\emph{Journal of Educational Measurement, 51}(3), 223-238.
}
\keyword{KernEqWPS}
