% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearEquate.R
\name{LinearEquate}
\alias{LinearEquate}
\title{Apply linear equating to data from a EG design.}
\usage{
LinearEquate(xscores, yscores)
}
\arguments{
\item{xscores}{A vector of scores from individual candidates on form X.}

\item{yscores}{A vector of scores from individual candidates on form Y.}
}
\value{
The function returns a list with the following elements:
\describe{
  \item{lys}{A function that translates any vector of scores on form X into equivalent scores on form Y.}
  \item{yx}{A vector representing the scores derived from applying yxFunc to a sorted list of all the unqiue form X scores that actually occur in the data.}
}
}
\description{
Apply linear equating to data from a EG design.
}
\examples{
#Simulate two vectors with differencein means of 10
x=rnorm(500,50,10)
y=rnorm(500,60,10)
lineq=LinearEquate(x,y)
lineq$lys(45:50)

}
