% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathsdata.R
\docType{data}
\name{mathsdata}
\alias{mathsdata}
\title{Simulated item score data}
\format{
A data frame with 2000 rows and 124 variables.
}
\usage{
mathsdata
}
\description{
A simulated data set of scores derived from IRT analysis of real maths items.
This data is provided purely for the purposes of demonstrating some of the functionality of this R package when applied to fairly realistic data.
}
\details{
A data set of 124 simulated item scores from a maths exam for 2000 pupils.
}
\keyword{datasets}
