% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIDIAweightingFUNC.R
\name{MDIAwtGroups}
\alias{MDIAwtGroups}
\title{Function to find weights to apply to a number of groups in a data set to ensure equivalence on some variables.}
\usage{
MDIAwtGroups(formula, data, groups, targetgroup = NA)
}
\arguments{
\item{formula}{A formula indicating the variables on which matching should be achieved after weighting. For example, ~x+I(x^2)+y+I(y^2)+I(y^3) would achieve matched values on the mean value of x, the squared mean value of x (essentially equivalent to the variance), and (essentially) the mean, standard deviation and skewness of y.}

\item{data, }{A data frame.}

\item{groups}{A vector indicating the group to which each row of the data belongs.}

\item{targetgroup}{Single value indicating which group defines the population to which all other groups should be weighted. By default all groups are weighted to match the overall mean.}
}
\value{
A vector of weights.
}
\description{
Method is from Haberman, 1984.
}
\references{
Haberman, S. J. (1984). Adjustment by minimum discriminant information. 
\emph{The Annals of Statistics, 12}(3), 971-988.
}
\keyword{KernEqWPS}
