% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel Equating Functions.R
\name{PEN1}
\alias{PEN1}
\title{Calulate PEN1 criterion for match of continuized and original score distributions}
\usage{
PEN1(h, scores, fX)
}
\arguments{
\item{h}{Bandwith being considered.}

\item{scores}{Vector of possible scores.}

\item{fX}{Number of people achieving each score (weighted numbers, probabilities or proportions also acceptable). Must be same length as scores.}
}
\value{
The function returns the value of PEN1 for the given bandwidth.
}
\description{
This function calculates the penalty term for a given bandwidth against a given distribution.
The definition of PEN1 can be found in Liang and von Davier (2014). The values of PEN1 are not of much interest in their own right.
Finding the bandwidth that minimises this penalty (using \emph{optPEN1}) can potentially be useful as identifying a plausible bandwidth to use in subsequent kernel equating.
}
\examples{
x1=rowSums(mathsdata)#make some scores
tab=tabulate(x1+1)#count how many people got each integer score
cbind(0:max(x1),tab)#look at the table
PEN1(0.5,0:max(x1),tab)#calculate PEN1 for a given bandwidth
optPEN1(0:max(x1),tab)#optimise PEN1
}
\references{
Liang, T., & von Davier, A. A. (2014). Cross-validation: An alternative bandwidth-selection method in kernel equating. 
\emph{Applied Psychological Measurement, 38}(4), 281-295.
}
\keyword{KernEqWPS}
