% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuckerEquate.R
\name{TuckerEquate}
\alias{TuckerEquate}
\title{Apply Tucker Observed Score equating to data from a NEAT design.}
\usage{
TuckerEquate(dx, dy, ws = NA)
}
\arguments{
\item{dx}{Data frame with variables "x" and "a" representing scores for individual candidates on form X and on the anchor test.}

\item{dy}{Data frame with variables "y" and "a" representing scores for individual candidates on form Y and on the anchor test.}

\item{ws}{Vector of two elements denoting the relative weights of the dx population and the dy population in the synthetic population where equating takes place.}
}
\value{
The function returns a list with the following elements:
\describe{
  \item{lys}{A function that translates any vector of scores on form X into equivalent scores on form Y.}
  \item{EqTable}{A data frame combining the sorted unique scores on form X in the data and their equated values on form Y.}
  \item{muSx}{Estimated mean on form X within the synthetic population.}
  \item{sigmaSx}{Estimated standard deviation on form X within the synthetic population.}
  \item{muSy}{Estimated mean on form Y within the synthetic population.}
  \item{sigmaSy}{Estimated standard deviation on form Y within the synthetic population.}
}
}
\description{
Apply Tucker Observed Score equating to data from a NEAT design.
}
\examples{
#Simulate two data sets with roughly equivalent relationship to underlying "true" scores but a difference in means
n1=8000
n2=5500
t1=rnorm(n1,0.5,1)
t2=rnorm(n2,0,1)
x=round(pmin(100,pmax(0,50+20*(0.9*t1+rnorm(n1,0,sqrt(1-0.9^2))))))
a1=round(pmin(20,pmax(0,10+4*(0.7*t1+rnorm(n1,0,sqrt(1-0.7^2))))))
cor(cbind(x,t1,a1))
y=round(pmin(100,pmax(0,50+20*(0.9*t2+rnorm(n2,0,sqrt(1-0.9^2))))))
a2=round(pmin(20,pmax(0,10+4*(0.7*t2+rnorm(n2,0,sqrt(1-0.7^2))))))
cor(cbind(y,t2,a2))
TuckerEquate(data.frame(x=x,a=a1),data.frame(y=y,a=a2))

}
